feat(types,backend-core): Consolidate Clerk issued JWT claims under ClerkJWTClaims
