+++ packages/backend-core/src/index.ts
@@ -6,4 +6,3 @@ export { AuthStatus } from './types';
 export type { ClerkFetcher } from './api/utils/RestClient';
 export type { Session } from './api/resources/Session';
 export type { Nullable } from './util/nullable';
-export type { JWTPayload } from './types';
